Fixed bug with methdo signature
	* Signature of the method doubleToByteBuffer was wrong, it
	  was accepting a long instead of a double

+++ backend/src/com/mymed/model/core/wrappers/cassandra/api07/MConverter.java
@@ -274,7 +274,7 @@ public final class MConverter {
 	 *            the double to convert
 	 * @return the ByteBuffer value
 	 */
-	public static ByteBuffer doubleToByteBuffer(final long value) {
+	public static ByteBuffer doubleToByteBuffer(final double value) {
 
 		final ByteBuffer buffer = ByteBuffer.allocate(DOUBLE_SIZE);
